#29 - Renamed new puzzle to "tower"
