+++ src/puzzle/tower/help.html
