+++ src/puzzle/tower/puzzle.html
