+++ src/puzzle/tower/puzzle.mjs
